CRUD for keys.
View keys in protected routes
